apps/graphs/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
langgraph-cli>=0.2.0
tavily-python>=0.5.0
langchain-community>=0.3.0
psycopg2-binary>=2.9.0